\name{histone_h3k4me1}
\alias{histone_h3k4me1}
\alias{h3k4me1}
\docType{data}
\title{
A toy dataset of H3k4me1 peaks to test within SMITE
}
\description{
A toy dataset of H3k4me1 peaks from liver ChIP-seq through the encode project.  
The file is a 
BED file.
}
\usage{data(histone_h3k4me1)}
\format{
  A data frame with 75448 observations on the following 3 variables.
  \describe{
    \item{column1}{a character vector specifying chromsome}
    \item{column2}{an integer vector specifying start}
    \item{column3}{an integer vector specifying end}
  }
}
\details{
This is a BED file that specifices the consensus locations of three H3K4me1 
ChIP-seq experiments performed on normal adult liver.
}
\value{
A dataframe in BED format (chromosome, start, end).
}
\source{
GSM669972, GSM621654, GSM537706

Roadmap Epigenomics
Lister R, et al. Nature. 2009 Nov 19;462(7271):315-22
ENCODE Project Consortium. An integrated encyclopedia of DNA elements in the 
human genome. Nature. 2012 Sep 6;489(7414):57-74.

}

\examples{
data(histone_h3k4me1)
head(h3k4me1)
}
\keyword{datasets}